bump

+++ ipv6er.sh
@@ -57,7 +57,14 @@ GPG_LOCATION="https://download.docker.com/linux/$KEYRING_DISTRO/gpg"
 DOCKER_KEYRING="/usr/share/keyrings/docker-archive-keyring.gpg"
 DOCKER_DOWNLOAD_LOCATION="https://download.docker.com/linux/$KEYRING_DISTRO"
 DOCKER_VERSION="stable"
-curl -fsSL "$GPG_LOCATION" | gpg --dearmor -o "$DOCKER_KEYRING"
+
+if test -f "$DOCKER_KEYRING"; then
+  echo "$DOCKER_KEYRING already exists"
+else
+  echo "$DOCKER_KEYRING does not exist"
+  echo "url -fsSL $GPG_LOCATION | gpg --dearmor -o $DOCKER_KEYRING"
+  curl -fsSL "$GPG_LOCATION" | gpg --dearmor -o "$DOCKER_KEYRING"
+fi
 
 echo "deb [arch=amd64 signed-by=$DOCKER_KEYRING] $DOCKER_DOWNLOAD_LOCATION $LSBCS $DOCKER_VERSION | tee $DOCKER_SOURCES_LIST > /dev/null"
 echo "deb [arch=amd64 signed-by=$DOCKER_KEYRING] $DOCKER_DOWNLOAD_LOCATION $LSBCS $DOCKER_VERSION"  | tee "$DOCKER_SOURCES_LIST" > /dev/null
